Undergraduate Research Software
Commit includes two screenshots from a beta build of the software. The software opened by asking the user for a UserID, sign a consent form, and complete a demographics survey. Then the user was presented with the experimental interface. The experimental interface presented the user with number on the left (keyboard stimuli) and right (mouse stimuli). The user would then input the key with the keyboard stimulus, and click on the screen button for the mouse stimulus. Data was exported via email port to a private data collection email. For questions, contact Preston Menke